Fix env handling
Source `rbenv-plugin-env` instead of trying to source `rbenv-path-env`,
which is not present in the `rbenv-plugin` install (and would introduce
a circular reference even if correctly pointing to `rbenv-path`, as the
recommended way to install `rbenv-path` is ... `rbenv-plugin`).

No replacement for `rbenv-path-shared` needed, as the functions defined
in that file are not used anywhere in the code base.